requirements.txt
atproto==0.0.50
certifi==2024.6.2
charset-normalizer==3.3.2
idna==3.7
numpy==2.0.0
oauthlib==3.2.2
pandas==2.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.1
requests==2.32.3
requests-oauthlib==1.3.1
setuptools==70.0.0
six==1.16.0
tzdata==2024.1
upgrade-requirements==1.7.0
urllib3==2.2.2
wheel==0.43.0
